New (complete) MockStorage in project UI.
The only missing bit is Copy, whose I don't know the real semantic.
I'm not sure about the presence of the personal list in the
different {All,Public}{Owned,Followed}Lists.
I moreover corrected a few uses of IListModelCollection, and translated
the code from French to English in a few places.